requirements.txt
langchain
openai
python-dotenv
streamlit
plotly
tiktoken
matplotlib
sentence-transformers
jupiter
pypdf
PyPDF2